Move "Pimcore\BootstrapPimcore" to the namespace root
