+++ tests/bootstrap.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use Neusta\Pimcore\TestingFramework\Pimcore\BootstrapPimcore;
+use Neusta\Pimcore\TestingFramework\BootstrapPimcore;
 use Neusta\Pimcore\TestingFramework\TestKernel;
 
 include dirname(__DIR__) . '/vendor/autoload.php';
